rename fallbackTexts file to fallbackContent, fix import in Section file
